examples fungible and non-fungible mint with more detail
